+++ app/Http/Controllers/Website/CompanyController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Website;
 
 use App\Http\Controllers\Controller;
 use App\Models;
+use App\Notifications\NewReviewSubmited;
 use Auth;
 use Carbon\Carbon;
 use File;
@@ -136,6 +137,9 @@ class CompanyController extends Controller
         $review->save();
 
         if ($review->id) {
+            $company = Models\Company::id($review->company_id)->first();
+            $review_data = Models\Review::id($review->id)->withBelongings()->first();
+            $company->notify(new NewReviewSubmited($review_data));
             return response()->json(['status' => 'success', 'message' => 'Review Stored'], 200);
         }