+++ resources/views/layouts/admin.blade.php
@@ -10,7 +10,7 @@
   <link rel="manifest" href="{{ asset('images/site.webmanifest') }}">
   <!-- Main Stylesheet File -->
   {{-- <link href="{{ mix('css/app.css') }}" rel="stylesheet"> --}}
-  <link href="{{ asset('css/app.css') }}" rel="stylesheet">
+  <link href="{{ mix('css/app.css') }}" rel="stylesheet">
  <link href="https://cdn.jsdelivr.net/npm/remixicon@2.5.0/fonts/remixicon.css" rel="stylesheet">
 
   <title>Admin</title>